+++ src/shared/query/query.types.ts
@@ -1,0 +1,9 @@
+import { TransactionFilters } from "./filtering.types.js";
+import { PaginationOptions } from "./pagination.types.js";
+import { TransactionSorting } from "./sorting.types.js";
+
+export interface TransactionQueryOptions {
+  pagination: PaginationOptions;
+  filters?: TransactionFilters;
+  sorting?: TransactionSorting;
+}