+++ src/shared/query/sorting.types.ts
@@ -1,0 +1,14 @@
+/*
+=========================================
+TRANSACTION SORTING
+=========================================
+*/
+
+export type transactionSortField = "amount" | "transactionDate" | "createdAt";
+
+export type SortOrder = "asc" | "desc";
+
+export interface TransactionSorting {
+  sortBy: transactionSortField;
+  sortOrder: SortOrder;
+}